- Implemented shader reflection (constant buffer only) - Reflection upgrades - Started working on UI

+++ Engine/Source/Rendering/DX12/DX12Statics.h
@@ -4,7 +4,7 @@
 
 namespace DX12Statics
 {
-    constexpr uint32 CalculateConstantBufferSize(uint32 byteSize)
+    constexpr uint64 CalculateConstantBufferSize(uint64 byteSize)
     {
         return (byteSize + 255) & ~255;
     }
